backend/lambda/image-to-pdf/requirements.txt
boto3==1.34.0
Pillow==10.1.0
requests==2.31.0
